chore: move EthereumBalanceProvider to separate file
